+++ app/__init__.py
@@ -2,6 +2,6 @@
 Pneumonia Detection API
 """
 
-__version__ = "3.4.2"
+__version__ = "3.6.0"
 __author__ = "IbnuSabillGitHub"
-__description__ = "Educational pneumonia detection API using deep learning"
+__description__ = "Educational pneumonia detection API"